src/requirements.txt
django==5.2.2
mysqlclient==2.2.7
Pillow==11.2.1
gunicorn==23.0.0
django-constance==4.3.2
redis==6.2.0
django-crispy-forms==2.4
crispy-bootstrap5==2025.4
django-cms==5.0.1
djangorestframework==3.16.0
django-filter==25.1